backend/requirements.txt
fastapi
uvicorn
spacy
sentence-transformers
scikit-learn
pandas
numpy
diffusers
transformers
torch
pillow
